* delete-command-bug:
  test: add delete patient appointment test case
  refactor: change variable name to camel casing
  docs: add javadoc for constructor method
  refactor: add additional parameters
  feat: add support for force delete of appointments when delete patient command is called
  refactor: change isPatientInAppointmentSchedule to hasPatientInAppointmentSchedule
  fix: change initial objects for test cases
  fix: resolve IndexOutOfBounds Exception
  docs: update DG with delete command use cases
  test: add isPatientInAppointmentSchedule method in ModelStub
  fix: change import ordering
  chore: group commands by appointments and patients
  refactor: shift MESSAGE_DELETE_PATIENT_SUCCESS to Messages class
  fix: add exception for deleting patients with existing appointments
  chore: change ObservableList to List

+++ src/main/java/seedu/address/logic/commands/patient/DeletePatientCommand.java
@@ -24,12 +24,23 @@ public class DeletePatientCommand extends Command {
             + "Parameters: INDEX (must be a positive integer)\n"
             + "Example: " + COMMAND_WORD + " 1";
 
-    public static final String MESSAGE_DELETE_PATIENT_SUCCESS = "Deleted Person: %1$s";
+    public static final String FORCE_DELETE_MESSAGE_USAGE = COMMAND_WORD + " --force"
+            + ": Deletes the person identified by the index number used in the displayed person list,\n"
+            + "along with all the existing appointments associated with the person in the appointment schedule.\n"
+            + "Parameters: INDEX (must be a positive integer)\n"
+            + "Example: " + COMMAND_WORD + " --force " + " 1";
 
     private final Index targetIndex;
+    private final boolean isForceDelete;
 
-    public DeletePatientCommand(Index targetIndex) {
+    /**
+     * Constructor: creates a DeletePatientCommand
+     * @param targetIndex index of patient to be deleted
+     * @param isForceDelete true if force delete is required
+     */
+    public DeletePatientCommand(Index targetIndex, boolean isForceDelete) {
         this.targetIndex = targetIndex;
+        this.isForceDelete = isForceDelete;
     }
 
     @Override
@@ -42,8 +53,19 @@ public class DeletePatientCommand extends Command {
         }
 
         Patient patientToDelete = lastShownList.get(targetIndex.getZeroBased());
+
+        if (isForceDelete) {
+            model.deletePatientAppointments(patientToDelete);
+        }
+
+        // checks if patient has any existing appointments
+        if (model.hasPatientInAppointmentSchedule(patientToDelete)) {
+            throw new CommandException(String.format(
+                    Messages.MESSAGE_FORCE_DELETE_REQUIRED, FORCE_DELETE_MESSAGE_USAGE));
+        }
+
         model.deletePatient(patientToDelete);
-        return new CommandResult(String.format(MESSAGE_DELETE_PATIENT_SUCCESS, patientToDelete));
+        return new CommandResult(String.format(Messages.MESSAGE_DELETE_PATIENT_SUCCESS, patientToDelete));
     }
 
     @Override
